perlmods: Simplify the AUTOLOAD code in NDF to stop it calling AutoLoader
Before if it could not find a function it would call AutoLoader which
would attempt to dynamically load it from some other file (eg .al).
This is never the right thing to do for the NDF module since it
does not actually try to dynamically load anything other than
Constants.