Add nullable PTROBJ typemap entry for HDSLoc *
This is a modified version of the standard T_PTROBJ typemap entry where
the output routine accepts a null pointer, packaging it in an object
as normal.  This allows the DAT__ROOT function to export a null pointer
which can then be passed to the C version of NDF.  Otherwise a null
pointer would be output as undef, which would be rejected by the
T_PTROBJ typemap input routine as not being of the expected class.

The alternative would be to instead alter the input routine to allow
undef as well as objects of the expected class.  That would allow
users of this module to simply write "undef" rather than "NDF::DAT__ROOT"
in the relevant NDF calls.